Merge to main

+++ account-service/src/main/java/com/achsat/crypto/account/repository/CashAccountRepository.java
@@ -1,11 +1,11 @@
 package com.achsat.crypto.account.repository;
 
-import com.achsat.crypto.account.model.CashAccount;
+import com.achsat.crypto.entity.model.CashAccount;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Mono;
 
 public interface CashAccountRepository extends ReactiveCrudRepository<CashAccount, Integer> {
 
     Mono<CashAccount> findCashAccountByAccountId(Integer id);
-    Mono<CashAccount> findCashAccountByCustomerId(Integer id);
+    Mono<CashAccount> findCashAccountByCustomerId(Long id);
 }
